Change concepts

+++ Source/Events.Processing/StreamPosition.cs
@@ -6,9 +6,9 @@ using Dolittle.Concepts;
 namespace Dolittle.Runtime.Events.Processing
 {
     /// <summary>
-    /// An unsigned long used to identify the position, or offset, of a stream.
+    /// An unsigned long used to identify the position of a stream.
     /// </summary>
-    public class StreamPosition : ConceptAs<ulong>
+    public class StreamPosition : ConceptAs<uint>
     {
         /// <summary>
         /// Represents the initial value of the <see cref="StreamPosition">position</see>.
@@ -19,13 +19,13 @@ namespace Dolittle.Runtime.Events.Processing
         /// Initializes a new instance of the <see cref="StreamPosition"/> class.
         /// </summary>
         /// <param name="position">Value representing the <see cref="StreamPosition"/>.</param>
-        public StreamPosition(ulong position) => Value = position;
+        public StreamPosition(uint position) => Value = position;
 
         /// <summary>
-        /// Implicitly convert from <see cref="ulong" /> to <see cref="StreamPosition" />.
+        /// Implicitly convert from <see cref="uint" /> to <see cref="StreamPosition" />.
         /// </summary>
-        /// <param name="position">Position number as <see cref="ulong"/>.</param>
-        public static implicit operator StreamPosition(ulong position) => new StreamPosition(position);
+        /// <param name="position">Position number as <see cref="uint"/>.</param>
+        public static implicit operator StreamPosition(uint position) => new StreamPosition(position);
 
         /// <summary>
         /// Increments the <see cref="StreamPosition" />.
